Set content type headers in responses
The responses that are sent from cropmail will always be in the UTF-8
character set (usually in the ASCII subset), so explicitly state that in
the Content-Type header.  Also state that the Content-Transfer-Encoding
is 8 bit since cropmail doesn't do encoding itself.

Previously receiving mail clients should have been assuming that the
responses were in ASCII, that will be true in most cases but won't be
when responding to an entry that included non-ASCII characters.  Now
those should be handled correctly.